+++ api/ssi-bridge/src/tools/migration/migration-to-0.1.3.ts
@@ -1,7 +1,8 @@
 import * as dotenv from 'dotenv';
-import { MongoDbService } from '@iota/is-shared-modules/node';
 
 dotenv.config();
+
+import { MongoDbService } from '@iota/is-shared-modules/node';
 
 const migrate = async () => {
 	console.log('starting migration...');