+++ api/ssi-bridge/CHANGELOG.md
@@ -11,3 +11,6 @@
 ### 0.1.6 (2022-07-05)
 
 - Updated dependencies
+- Reduce jwt overhead
+- Add Manager role
+- Implement hidden identities